[Refactor/TF] Refactor tf tensor-filter subplugin to remove .c/.cc wrapper
This commit refactors tf tensor-filter subplugin to remove .c/.cc wrapper.
